Retry on errors that indicate cluster overloading.
When PUT-ting an object into Swift, we may encounter a 503; this seems
to indicate that the cluster is overloaded. We will retry on
encountering a 503 (up to 10 times). This way, a momentary overloading
of Swift doesn't cause benchmark workers to die. However, the workers
won't report completion of the upload until after a successful upload.

Similarly, when DELETE-ing an object, we ignore 404s and 503s. 404s
can occur if nodes A and B received the upload, but node C hasn't
gotten it yet. Then, when a DELETE request comes in, node C gets the
request and returns file-not-found. Since we know the file is there
(we put it there), we retry until the DELETE succeeds.