Add stats output.
Currently, master.py prints a CSV on stdout that has a set of
(unixtime, files-uploaded) pairs. The times are truncated to the
nearest second. Eyeballing a graph of this should give you some idea
based on the line height of the object-ingest rate of the cluster.

Also, tweak retry count up to 50 from 10 so that DELETEs are less
likely to blow up.