Let the master run for more than two minutes.
In beanstalkd land, a job's lifecycle looks like this:

   put with delay               release with delay
  ----------------> [DELAYED] <------------.
                        |                   | '-` touch (reset TTR)
                        | (time passes)     | | |
                        |                   | | |
   put                  v     reserve       | | v   delete
  -----------------> [READY] ---------> [RESERVED] --------> *poof*
                       ^  ^                | |  |
                       |  |\  release      | |  |
                       |  | `-------------'  |  |
                       |  |                  |  |
                       |  \     times out    |  |
                       |   `----------------'   |
                       |                        |
                       | kick                   |
                       |                        |
                       |       bury             |
                    [BURIED] <-----------------'
                       |
                       |  delete
                        `--------> *poof*

In particular, notice how jobs will time out if they're reserved but
not deleted, buried, released, or touched.

beanstalkd has a feature where, if you
  1. have a job that is about to time out
  2. try to reserve a job
  3. would block on reserving that job due to empty queue(s)
then instead of blocking, it replies with a DEADLINE_SOON message.

The goal is to prevent you from letting your active job(s) time out
while you sit waiting for more data.

The default TTR (job timeout) is 120 seconds.

The symptom of this bug was that the master would put jobs into the
queue, then read stats for two minutes, then die with a DeadlineSoon
exception. We were attempting to reserve a job and then delete it
immediately, but that wasn't happening because, in Python,
`job.delete` and `job.delete()` are not the same at all.

It's hard to see on small job sets because any bunch of work that
completed in under 120 seconds wouldn't trigger the bug. It's only on
larger benchmarking runs that this pops up.

Python is hard. Let's go shopping.