Use one container per 1000 objects in object-creation benchmark.
The goal here is to avoid overwhelming any single container with
object-creation requests. Since each container has its own sqlite
database, it's entirely possible for the database to be a bottleneck
instead of the cluster's disks. This makes Swift look bad.

The recommended way to use Swift is to split up your objects between
containers a bit; don't overload any one container.

The number 1000 is just something we made up. If you're thinking of
changing it to make it seem less arbitrary, I'd suggest either 1024 or
1023; powers of 2 make people think there are hidden fixed-width
reasons for a number. ;)

You can override the default container count with the command-line
option -C <n> where n is the number of containers that you want. Note
that this is the total number of containers to use, not the number of
objects to put in a particular container.