Fixes #1: handle zero responses per CRUD type
Drive-by fix for writing/reading the pickled stats data in "binary"
mode.  A saved stats run presumably generated on a Windows server was
unable to be loaded on OS X because the pickled file had \r\n in it
instead of just \n.  (This was worked around by replacing \r\n with \n
in a hex editor.)