Added reporting for per-size CRUD.
Added a "size" section to the report which now includes each size
category specified in the scenario file.  For each size category, we
include the percentage of total operations which will utilize that size,
the CRUD profile for that size, and the range of bytes for the size.
After all the sizes, we list the overall, weighted average CRUD
percentages.

Made everything robust to having a size category defined in the scenario
which is not included in the initial_files dictionary.  (They will get
no jobs but will appear in the report's size section with a "0" in the
"% Ops" column.)

Renamed "Swift TX ID for worst latency" column to "Worst latency TX ID".