Add output for the number of errors and retries
This patch enables ssbench to show # of errors and retries
for availability calculation.
This patch enables to show # of errors and retries and
retry rate in addition to current output.

Each item of additional output means as follows:

Errors: The sum of errors which occur during benchmarking.
Retries: The sum of retries for each job.
         (Count up by worker when catching ignoring http
          errors or socket errors)
Retry Rate: The proportion of retries. (# of retries / # of requests)
            i.e. The rate may be higher than 100 %.

Additionally, semantics of 'Count' is changed
to include # of errors.
(Previously, ssbench excluded errors from 'Count')

[Output Example]
For a run which attempted 10 READ operations, with one Exception
with 10 retries, 9 successes with 4 retries.

'''
  READ
         Count:    10 (    1 error;     14 retries:  140.00%)  Average requests per second:  88.9
'''