Allow ssbench-master to override worker block_size
Allow ssbench-master to override the default work block size used for
GET/PUT operations (-b BYTES or --block-size BYTES)

Always print the distribution of requests per worker-ID in the report.

Document the container_base key in the scenario file and allow a
non-default value for that to be passed into the cleanup-containers
sub-command.

Bump version to 0.2.22.