add delete-after option
To emulates continuous loads of CREATE with X-Delete-After header,
this patch adds delete-after option for scenario file and args of
ssbench-master run-scenario.

A delete_after value(to be set integer as seconds) appends
expiring time to all objects.
If setting 0 (or None by default), this feature is disable and all
objects will not be expired. This value may be overridden for any
given run with the --delete-after DELETE_AFTER flag to ssbench-master
run-scenario.