allow seting the container policy
Set it with -p/--policy on the CLI or set it by using "policy" in
the scenario file.

The containers created by ssbench will now have the policy name
appended to them. If no policy name is given (resulting in the
cluster default being used), the policy suffix is "default_policy".