Remove the dependency on statlib.
The statlib project is no longer actively developed. It is causing
issues when retrieving dependencies. We only rely on median and
finite population standard deviation computations from statlib. Given
the above factors, we can remove the statlib dependency and implement
the required functions ourselves.

The commit adds the required stats functions in ssbench/util.py and
supplies tests for them. The README instructions are updated to remove
references to the dependency and the requirements.txt is modified,  as
well.